reorganize elements
