+++ project/config/packages/doctrine.yaml
@@ -10,9 +10,15 @@ doctrine:
         naming_strategy: doctrine.orm.naming_strategy.underscore_number_aware
         auto_mapping: true
         mappings:
-            App:
+            Player:
                 is_bundle: false
                 type: annotation
-                dir: '%kernel.project_dir%/src/Entity'
-                prefix: 'App\Entity'
-                alias: App
+                dir: '%kernel.project_dir%/src/Players/Entity'
+                prefix: 'App\Players\Entity'
+                alias: Players
+            Team:
+                is_bundle: false
+                type: annotation
+                dir: '%kernel.project_dir%/src/Teams/Entity'
+                prefix: 'App\Teams\Entity'
+                alias: Teams
